refactor: merge processor into the BlockEventListener.
